# docs/swagger.yaml
definitions:
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /friends/{friend_id}:
    delete:
      description: Remove an existing friend from your friends list
      parameters:
      - description: Friend's user ID
        in: path
        name: friend_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Friend removed
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Remove friend
      tags:
      - friends
  /friends/{friend_id}/accept:
    post:
      description: Accept a pending friend request from another user
      parameters:
      - description: Friend's user ID
        in: path
        name: friend_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Friend request accepted
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Accept friend request
      tags:
      - friends
  /friends/{friend_id}/decline:
    post:
      description: Decline or cancel a pending friend request
      parameters:
      - description: Friend's user ID
        in: path
        name: friend_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Friend request declined
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Decline friend request
      tags:
      - friends
  /friends/add:
    post:
      consumes:
      - application/json
      description: Send a friend request to another user by their user ID
      parameters:
      - description: 'Friend request — {friend_id: string}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Friend request sent
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Cannot add yourself
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Send friend request
      tags:
      - friends
...
  /users/friends:
    get:
      description: Get the authenticated user's friends list with pagination
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Friends list
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List friends
      tags:
      - friends
  /users/friends/pending:
    get:
      description: Get all incoming friend requests awaiting acceptance
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Pending requests
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List pending friend requests
      tags:
      - friends
...
swagger: "2.0"